feat: add reusable starfield component
